+++ configs/recognition/tsn/tsn_r50_1x1x8_50e_hmdb51_imagenet_rgb.py
@@ -1,45 +1,35 @@
-# model settings
 model = dict(
     type='Recognizer2D',
     backbone=dict(
-        type='ResNetTSM',
+        type='ResNet',
         pretrained='torchvision://resnet50',
         depth=50,
-        norm_eval=False,
-        shift_div=8),
+        norm_eval=False),
     cls_head=dict(
-        type='TSMHead',
-        num_classes=174,
+        type='TSNHead',
+        num_classes=51,
         in_channels=2048,
         spatial_type='avg',
         consensus=dict(type='AvgConsensus', dim=1),
-        dropout_ratio=0.5,
-        init_std=0.001,
-        is_shift=True))
-# model training and testing settings
+        dropout_ratio=0.4,
+        init_std=0.01))
 train_cfg = None
-test_cfg = dict(average_clips='prob')
-# dataset settings
+test_cfg = dict(average_clips=None)
 dataset_type = 'RawframeDataset'
-data_root = 'data/sthv2/rawframes'
-data_root_val = 'data/sthv2/rawframes'
-ann_file_train = 'data/sthv2/sthv2_train_list_rawframes.txt'
-ann_file_val = 'data/sthv2/sthv2_val_list_rawframes.txt'
-ann_file_test = 'data/sthv2/sthv2_val_list_rawframes.txt'
+data_root = 'data/hmdb51/rawframes'
+data_root_val = 'data/hmdb51/rawframes'
+ann_file_train = 'data/hmdb51/hmdb51_train_split_1_rawframes.txt'
+ann_file_val = 'data/hmdb51/hmdb51_val_split_1_rawframes.txt'
+ann_file_test = 'data/hmdb51/hmdb51_val_split_1_rawframes.txt'
 img_norm_cfg = dict(
     mean=[123.675, 116.28, 103.53], std=[58.395, 57.12, 57.375], to_bgr=False)
 train_pipeline = [
     dict(type='SampleFrames', clip_len=1, frame_interval=1, num_clips=8),
     dict(type='RawFrameDecode'),
     dict(type='Resize', scale=(-1, 256)),
-    dict(
-        type='MultiScaleCrop',
-        input_size=224,
-        scales=(1, 0.875, 0.75, 0.66),
-        random_crop=False,
-        max_wh_scale_gap=1,
-        num_fixed_crops=13),
+    dict(type='RandomResizedCrop'),
     dict(type='Resize', scale=(224, 224), keep_ratio=False),
+    dict(type='Flip', flip_ratio=0.5),
     dict(type='Normalize', **img_norm_cfg),
     dict(type='FormatShape', input_format='NCHW'),
     dict(type='Collect', keys=['imgs', 'label'], meta_keys=[]),
@@ -50,11 +40,12 @@ val_pipeline = [
         type='SampleFrames',
         clip_len=1,
         frame_interval=1,
-        num_clips=16,
+        num_clips=8,
         test_mode=True),
     dict(type='RawFrameDecode'),
     dict(type='Resize', scale=(-1, 256)),
-    dict(type='CenterCrop', crop_size=224),
+    dict(type='CenterCrop', crop_size=256),
+    dict(type='Flip', flip_ratio=0),
     dict(type='Normalize', **img_norm_cfg),
     dict(type='FormatShape', input_format='NCHW'),
     dict(type='Collect', keys=['imgs', 'label'], meta_keys=[]),
@@ -65,18 +56,19 @@ test_pipeline = [
         type='SampleFrames',
         clip_len=1,
         frame_interval=1,
-        num_clips=16,
+        num_clips=8,
         test_mode=True),
     dict(type='RawFrameDecode'),
     dict(type='Resize', scale=(-1, 256)),
-    dict(type='CenterCrop', crop_size=224),
+    dict(type='CenterCrop', crop_size=256),
+    dict(type='Flip', flip_ratio=0),
     dict(type='Normalize', **img_norm_cfg),
     dict(type='FormatShape', input_format='NCHW'),
     dict(type='Collect', keys=['imgs', 'label'], meta_keys=[]),
     dict(type='ToTensor', keys=['imgs'])
 ]
 data = dict(
-    videos_per_gpu=6,
+    videos_per_gpu=32,
     workers_per_gpu=4,
     train=dict(
         type=dataset_type,
@@ -90,34 +82,26 @@ data = dict(
         pipeline=val_pipeline),
     test=dict(
         type=dataset_type,
-        ann_file=ann_file_test,
+        ann_file=ann_file_val,
         data_prefix=data_root_val,
         pipeline=test_pipeline))
-# optimizer
-optimizer = dict(
-    type='SGD',
-    constructor='TSMOptimizerConstructor',
-    paramwise_cfg=dict(fc_lr5=True),
-    lr=0.0075,  # this lr is used for 8 gpus
-    momentum=0.9,
-    weight_decay=0.0005)
-optimizer_config = dict(grad_clip=dict(max_norm=20, norm_type=2))
-# learning policy
+optimizer = dict(type='SGD', lr=0.025, momentum=0.9, weight_decay=0.0001)
+optimizer_config = dict(grad_clip=dict(max_norm=40, norm_type=2))
 lr_config = dict(policy='step', step=[20, 40])
 total_epochs = 50
-checkpoint_config = dict(interval=1)
+checkpoint_config = dict(interval=5)
 evaluation = dict(
-    interval=2, metrics=['top_k_accuracy', 'mean_class_accuracy'])
+    interval=1, metrics=['top_k_accuracy', 'mean_class_accuracy'], topk=(1, 5))
 log_config = dict(
     interval=20,
     hooks=[
         dict(type='TextLoggerHook'),
-        # dict(type='TensorboardLoggerHook'),
+        #    dict(type='TensorboardLoggerHook')
     ])
-# runtime settings
 dist_params = dict(backend='nccl')
 log_level = 'INFO'
-work_dir = './work_dirs/tsm_r50_1x1x8_50e_sthv2_rgb/'
+work_dir = './work_dirs/tsn_r50_1x1x8_50e_hmdb51_imagenet_rgb/'
 load_from = None
 resume_from = None
 workflow = [('train', 1)]
+gpu_ids = range(0, 1)