+++ mmaction/models/recognizers/recognizer2d.py
@@ -32,12 +32,9 @@ class Recognizer2D(BaseRecognizer):
 
         return losses
 
-    def forward_test(self, imgs):
-        """Defines the computation performed at every call when evaluation and
-        testing."""
-        test_crops = self.test_cfg.get('test_crops', None)
-        twice_sample = self.test_cfg.get('twice_sample', False)
-
+    def _do_test(self, imgs):
+        """Defines the computation performed at every call when evaluation,
+        testing and gradcam."""
         batches = imgs.shape[0]
 
         imgs = imgs.reshape((-1, ) + imgs.shape[2:])
@@ -57,13 +54,26 @@ class Recognizer2D(BaseRecognizer):
             losses.update(loss_aux)
             num_segs = 1
 
+        # When using `TSNHead` or `TPNHead`, shape is [batch_size, num_classes]
+        # When using `TSMHead`, shape is [batch_size * num_crops, num_classes]
+        # `num_crops` is calculated by:
+        #   1) `twice_sample` in `SampleFrames`
+        #   2) `num_sample_positions` in `DenseSampleFrames`
+        #   3) `ThreeCrop/TenCrop/MultiGroupCrop` in `test_pipeline`
+        #   4) `num_clips` in `SampleFrames` or its subclass if `clip_len != 1`
         cls_score = self.cls_head(x, num_segs)
-        if test_crops is not None:
-            if twice_sample:
-                test_crops = test_crops * 2
-            cls_score = self.average_clip(cls_score, test_crops)
 
-        return cls_score.cpu().numpy()
+        assert cls_score.size()[0] % batches == 0
+        # calculate num_crops automatically
+        cls_score = self.average_clip(cls_score,
+                                      cls_score.size()[0] // batches)
+
+        return cls_score
+
+    def forward_test(self, imgs):
+        """Defines the computation performed at every call when evaluation and
+        testing."""
+        return self._do_test(imgs).cpu().numpy()
 
     def forward_dummy(self, imgs):
         """Used for computing network FLOPs.
@@ -87,32 +97,4 @@ class Recognizer2D(BaseRecognizer):
     def forward_gradcam(self, imgs):
         """Defines the computation performed at every call when using gradcam
         utils."""
-        test_crops = self.test_cfg.get('test_crops', None)
-        twice_sample = self.test_cfg.get('twice_sample', False)
-
-        batches = imgs.shape[0]
-
-        imgs = imgs.reshape((-1, ) + imgs.shape[2:])
-        num_segs = imgs.shape[0] // batches
-
-        losses = dict()
-
-        x = self.extract_feat(imgs)
-        if hasattr(self, 'neck'):
-            x = [
-                each.reshape((-1, num_segs) +
-                             each.shape[1:]).transpose(1, 2).contiguous()
-                for each in x
-            ]
-            x, loss_aux = self.neck(x)
-            x = x.squeeze(2)
-            losses.update(loss_aux)
-            num_segs = 1
-
-        cls_score = self.cls_head(x, num_segs)
-        if test_crops is not None:
-            if twice_sample:
-                test_crops = test_crops * 2
-            cls_score = self.average_clip(cls_score, test_crops)
-
-        return cls_score
+        return self._do_test(imgs)